backend: Refactor reading of existing yaml config in make first-run
