+++ src/backend/cli/utils.py
@@ -1,4 +1,15 @@
-from backend.cli.constants import WELCOME_MESSAGE, DeploymentName, bcolors
+import sys
+from pathlib import Path
+
+import yaml
+
+from backend.cli.constants import (
+    ENV_YAML_CONFIG_MAPPING,
+    WELCOME_MESSAGE,
+    DeploymentName,
+    bcolors,
+)
+from backend.cli.setters import read_yaml
 
 
 def print_styled(text: str, color: str = bcolors.ENDC):
@@ -10,6 +21,43 @@ def show_welcome_message():
     print_styled(
         "👋 First things first, let's set up your environment.", bcolors.MAGENTA
     )
+
+
+def process_existing_yaml_config(secrets, path, prompt):
+    _path = Path(path)
+
+    if _path.is_file():
+        try:
+            yaml_config = read_yaml(path)
+        except yaml.scanner.ScannerError:
+            if prompt():
+                yaml_config = {}
+                _path.unlink()
+            else:
+                sys.exit(1)
+
+        secrets.update(convert_yaml_to_secrets(yaml_config))
+
+
+def convert_yaml_to_secrets(yaml_dict: dict):
+    def get_nested_value(d, path):
+        keys = path.split(".")
+        value = d
+        for key in keys:
+            if not isinstance(value, dict) or key not in value:
+                return None  # Return None if the path does not exist
+            value = value[key]
+        return value
+
+    secrets = {}
+    for env_var, mapping in ENV_YAML_CONFIG_MAPPING.items():
+        path = mapping.get("path")
+        if path:  # Only process mappings with a defined path
+            value = get_nested_value(yaml_dict, path)
+            if value is not None:  # Add only if the value exists
+                secrets[env_var] = value
+
+    return secrets
 
 
 def wrap_up(deployments):
